# openapi.yaml
openapi: 3.1.0
info:
  title: Family Hub API
  version: "1.0.0"
  description: |
    Family Hub backend API (Express + TypeScript + Supabase/Postgres).

    This spec documents the API **as it actually behaves today**, including
    two genuinely different response-envelope conventions and three
    different auth conventions, depending on which route file an endpoint
    lives in. See `docs/API.md` for a human-oriented companion/index to this
    spec, including narrative explanation of the quirks below.

    ## Response envelopes (two incompatible conventions)

    - **Newer routes** (`chores`, `learning`, `announcements`, `reminders`,
      `energy`, `calendar`, `family`, `activity-log`, `smartthings`) return
      an object with a top-level `status: 'success' | 'error'` field and a
      `timestamp` ISO string on every response. On error, the shape is
      `{ status: 'error', message: string, error?: string }`
      (see `EnvelopeError`). On success, most — but **not all** — of these
      routes nest the payload under a `data` key alongside `count?`
      (see `EnvelopeSuccess`). Several endpoints in this same "newer" group
      (`chores`, `smartthings`, `learning`, `performance`-adjacent routes)
      instead use an endpoint-specific top-level key (e.g. `chores`,
      `devices`, `stats`, `progress`, `activity`, `history`) instead of
      `data`. Each such endpoint documents its literal response shape
...
      path.
...

    ## Auth conventions (three, genuinely different)
...
servers:
...
security: []
paths:
...
  /api/kiosk/enroll:
    post:
      tags: [Kiosk]
      summary: Register this device as a family display (parent only)
      operationId: enrollKioskDevice
      security: [{ userIdHeader: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string, maxLength: 80, description: A name for the display, e.g. "Kitchen wall". }
      responses:
        '201':
          description: Device enrolled. The raw token is returned exactly once.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: success }
                  token: { type: string, description: Raw device token — store it on the display; not retrievable again. }
                  deviceId: { type: string, format: uuid }
                  familyId: { type: string, format: uuid }
                  timestamp: { type: string, format: date-time }
        '401': { description: Missing x-user-id., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '403': { description: Caller is not a parent/admin., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '500': { description: Unexpected failure., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
  /api/kiosk/devices:
    get:
      tags: [Kiosk]
      summary: List the household's registered displays (parent only)
      operationId: listKioskDevices
      security: [{ userIdHeader: [] }]
      responses:
        '200':
          description: The registered displays (no tokens).
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: success }
                  devices: { type: array, items: { $ref: '#/components/schemas/KioskDevice' } }
                  count: { type: integer }
                  timestamp: { type: string, format: date-time }
        '401': { description: Missing x-user-id., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '403': { description: Caller is not a parent/admin., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '500': { description: Unexpected failure., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
  /api/kiosk/devices/{id}:
    delete:
      tags: [Kiosk]
      summary: Revoke a registered display (parent only)
      operationId: revokeKioskDevice
      security: [{ userIdHeader: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Display revoked., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeSuccess' } } } }
        '401': { description: Missing x-user-id., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '403': { description: Caller is not a parent/admin., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '404': { description: No such display in this household., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '500': { description: Unexpected failure., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
  /api/kiosk/session:
    post:
      tags: [Kiosk]
      summary: Exchange a device token for the household boot payload (no user login)
      operationId: startKioskSession
      security: [{ kioskTokenHeader: [] }]
      responses:
        '200':
          description: The household, its members, and kiosk settings.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/KioskBootstrap'
                  - type: object
                    properties:
                      status: { type: string, const: success }
                      timestamp: { type: string, format: date-time }
        '401': { description: Missing or unrecognised x-kiosk-token., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '404': { description: Household not found., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '500': { description: Unexpected failure., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
  /api/kiosk/pin:
    put:
      tags: [Kiosk]
      summary: Set or replace the household PIN (parent only)
      operationId: setKioskPin
      security: [{ userIdHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties:
                pin: { type: string, pattern: '^\d{4}$', description: Exactly 4 digits. }
      responses:
        '200': { description: PIN set., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeSuccess' } } } }
        '400': { description: PIN is not exactly 4 digits., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '401': { description: Missing x-user-id., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '403': { description: Caller is not a parent/admin., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '500': { description: Unexpected failure., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
    delete:
      tags: [Kiosk]
      summary: Remove the household PIN (parent only)
      operationId: clearKioskPin
      security: [{ userIdHeader: [] }]
      responses:
        '200': { description: PIN removed., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeSuccess' } } } }
        '401': { description: Missing x-user-id., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '403': { description: Caller is not a parent/admin., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '500': { description: Unexpected failure., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
  /api/kiosk/pin/verify:
    post:
      tags: [Kiosk]
      summary: Check a PIN attempt (any member; rate-limited per device)
      operationId: verifyKioskPin
      security: [{ userIdHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pin]
              properties:
                pin: { type: string }
      responses:
        '200':
          description: PIN correct.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: success }
                  ok: { type: boolean, const: true }
                  timestamp: { type: string, format: date-time }
        '400': { description: pin missing., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '401': { description: Missing x-user-id, or incorrect PIN., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '429': { description: Too many attempts from this device., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }
        '500': { description: Unexpected failure., content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeError' } } } }

  # ================================================================
  # LEARNING (x-user-id)
  # ================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <token>` must be present. Used by `auth`,
        `users`, `badges`, `points`, `external-apis`. For most routes in
        this group the server only checks the header is present and
        prefixed `Bearer ` — it does not cryptographically verify the
        token (the one exception is `GET /auth/me`, which does call
        `supabase.auth.getUser(token)`).
    userIdHeader:
      type: apiKey
      in: header
      name: x-user-id
      description: |
        Custom header identifying the caller by user ID. Used by `chores`,
        `learning`, `announcements`, `reminders`, `energy`, `calendar`,
        `family`, `activity-log`. There is no cryptographic verification —
        any UUID-shaped (or even arbitrary) value is trusted as-is. A few
        endpoints within these route files (e.g. `GET /api/energy/usage`,
        `POST /api/activity/log`, several PATCH/DELETE-by-id endpoints in
        `announcements`/`reminders`/`calendar`) do not actually check this
        header despite living in the same route file — see the per-endpoint
        `security` field and description for exceptions.
    kioskTokenHeader:
      type: apiKey
      in: header
      name: x-kiosk-token
      description: |
        Raw device token issued by `POST /api/kiosk/enroll`, presented by a
        registered wall display on `POST /api/kiosk/session` to boot into its
        household with no user login. The server sha256-hashes it and matches
        against `kiosk_devices.token_hash`.
...
  schemas:
...
    EnvelopeSuccess:
      type: object
      description: |
        The success shape used by MOST (not all) "newer" routes. `count`
        is present on list endpoints, absent on single-resource endpoints.
        `data`'s actual type varies per endpoint — see the `allOf` override
        on each path.
      required: [status, timestamp]
      properties:
        status: { type: string, const: success }
        data: {}
        count: { type: integer }
        timestamp: { type: string, format: date-time }
    EnvelopeError:
      type: object
      description: The error shape used consistently by all "newer" routes (chores, learning, announcements, reminders, energy, calendar, family, activity-log, smartthings).
      required: [status, message]
      properties:
        status: { type: string, const: error }
        message: { type: string }
        error: { type: string, description: "Present on most (not all) error responses; usually a raw error message string." }
...
    KioskDevice:
      type: object
      description: A registered wall display, safe to list back to a parent (never the token).
      required: [id, label, lastSeenAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        lastSeenAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    KioskProfile:
      type: object
      description: One family member as shown on the kiosk profile picker.
      required: [userId, name, role, color]
      properties:
        userId: { type: string, format: uuid }
        name: { type: string }
        role: { type: string }
        color: { type: [string, 'null'] }
    KioskBootstrap:
      type: object
      description: The payload a wall display boots with — no user login involved.
      required: [family, members, hasPin, idleMinutes]
      properties:
        family:
          type: object
          required: [id, name]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
        members: { type: array, items: { $ref: '#/components/schemas/KioskProfile' } }
        hasPin: { type: boolean }
        idleMinutes: { type: integer }

    # ---------------- Learning (hand-rolled schema) ----------------
...